reconcil-invoices-bankstatements/requirements.txt
pandas>=2.0.0
pdfplumber>=0.9.0
PyMuPDF>=1.23.8
Pillow>=10.0.0
openai>=1.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0